Site Terms + Split-Back Photo Fix
Site Terms.pdf now in main directory with links at the footer of every
page to open the document in a new tab as a .pdf in browser. The photos
for the split-back photos are also changed to display the backs of the
panties as the default (as per Kris request).